pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ha-automation"
version = "1.0.0"
description = "Python CLI and library for managing Home Assistant automations"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    { name = "Home Assistant Automation" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
]

[project.scripts]
ha-automation = "ha_automation.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/ha-automation"
Issues = "https://github.com/yourusername/ha-automation/issues"

[tool.setuptools]
packages = ["ha_automation"]
